Tidied up the PR in response to code review
