+++ src/main/kotlin/com/romainpiel/catsanddogs/repository/ValuesMap.kt
@@ -1,13 +1,11 @@
-package com.romainpiel.catsanddogs
+package com.romainpiel.catsanddogs.repository
 
 import org.jetbrains.ktor.util.ValuesMap
 import java.time.OffsetDateTime
 
-
 fun ValuesMap.getDate(key: String): OffsetDateTime {
     val dateStr = get(key)
     val date = dateStr?.let { OffsetDateTime.parse(it) }
-    val safeFrom = date ?: OffsetDateTime.MIN
 
-    return safeFrom
+    return date ?: OffsetDateTime.MIN
 }
